flame and dragon 2.1

+++ 2D_Game_Engine/ECS/ProjectileComponent.h
@@ -8,7 +8,7 @@
 class ProjectileComponent : public Component
 {
 public:
-    ProjectileComponent(int rng, int sp, Vector2D vel, int dmg)
+    ProjectileComponent(int rng, float sp, Vector2D vel, int dmg)
     : range(rng), speed(sp), velocity(vel), damage(dmg)
     {}
     
@@ -23,7 +23,13 @@ public:
         std::unique_ptr<Vector2D> spriteSize(entity->getComponent<SpriteComponent>().getSize());
         transform->position.x -= spriteSize->x / 2;
         transform->position.y -= spriteSize->y / 2;
+        
     }
+    
+//    TransformComponent &getTransform()
+//    {
+//        return *transform;
+//    }
     
     void update() override
     {
@@ -39,7 +45,7 @@ private:
     TransformComponent *transform;
     Vector2D velocity;
     int range = 0;
-    int speed = 0;
+    float speed = 0;
     int distance = 0;
     int damage = 0;
 };
